Developing TraitsStep
